[SD-852] move controller to api namespace
